+++ lib/datacite/mapping/date.rb
@@ -87,7 +87,7 @@ module Datacite
 
       def <=>(other)
         return nil unless other.class == self.class
-        [:date_value, :range_start, :range_end].each do |v|
+        [:date_value, :range_start, :range_end, :type].each do |v|
           order = send(v) <=> other.send(v)
           return order if order.nonzero?
         end
@@ -95,7 +95,7 @@ module Datacite
       end
 
       def hash
-        [date_value, range_start, range_end].hash
+        [date_value, range_start, range_end, type].hash
       end
 
       def to_s